+++ coaster/sqlalchemy.py
@@ -28,6 +28,12 @@ class IdMixin(object):
     id = Column(Integer, primary_key=True)
 
 
+timestamp_columns = (
+    Column('created_at', DateTime, default=datetime.utcnow, nullable=False),
+    Column('updated_at', DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False),
+    )
+
+
 class TimestampMixin(object):
     """
     Provides the :attr:`created_at` and :attr:`updated_at` audit timestamps
@@ -119,7 +125,7 @@ class BaseScopedNameMixin(BaseMixin):
             organizer_id = db.Column(db.Integer, db.ForeignKey('organizer.id'))
             organizer = db.relationship(Organizer)
             parent = db.synonym('organizer')
-            __table_args__ = (db.UniqueConstraint('name', 'organizer_id'),)
+            __table_args__ = (db.UniqueConstraint('organizer_id', 'name'),)
     """
     @declared_attr
     def name(cls):
@@ -223,7 +229,7 @@ class BaseScopedIdMixin(BaseMixin):
             event_id = db.Column(Integer, db.ForeignKey('event.id'))
             event = db.relationship(Event)
             parent = db.synonym('event')
-            __table_args__ = (db.UniqueConstraint('url_id', 'event_id'),)
+            __table_args__ = (db.UniqueConstraint('event_id', 'url_id'),)
     """
     @declared_attr
     def url_id(cls):
@@ -266,7 +272,7 @@ class BaseScopedIdNameMixin(BaseScopedIdMixin):
             organizer_id = db.Column(db.Integer, db.ForeignKey('organizer.id'))
             organizer = db.relationship(Organizer)
             parent = db.synonym('organizer')
-            __table_args__ = (db.UniqueConstraint('url_id', 'organizer_id'),)
+            __table_args__ = (db.UniqueConstraint('organizer_id', 'url_id'),)
     """
     @declared_attr
     def name(cls):